Address slotnum review feedback
